promotion of float on instantiation.

+++ stwcs/distortion/models.py
@@ -244,7 +244,7 @@ class IDCModel(GeometryModel):
         if 'empty_model' in self.refpix and self.refpix['empty_model']:
             pass
         else:
-            self.refpix['PSCALE'] = np.float64(self.refpix['PSCALE']) * binned
+            self.refpix['PSCALE'] = self.refpix['PSCALE'] * binned
             self.cx = self.cx * binned
             self.cy = self.cy * binned
             self.refpix['XREF'] = self.refpix['XREF'] / binned
